Add welcoming settings and update localization strings
A new welcoming settings option has been added to the protection section of the dashboard sidebar to manage the welcome and leave messages for a server. The texts in some components have also been translated from Czech to English to improve the user interface and accessibility to non-Czech speakers. Also removed references to 'Anti-Raid' functionality due its redundancy.